Clean up viewables into separate files outside of scene
