+++ public/app/features/canvas/runtime/DimensionViewable.tsx
@@ -1,0 +1,33 @@
+import { MoveableManagerInterface, Renderer } from 'moveable';
+
+export const dimensionViewable = {
+  name: 'dimensionViewable',
+  props: {},
+  events: {},
+  render(moveable: MoveableManagerInterface<any, any>, React: Renderer) {
+    const rect = moveable.getRect();
+    return (
+      <div
+        key={'dimension-viewable'}
+        className={'moveable-dimension'}
+        style={{
+          position: 'absolute',
+          left: `${rect.width / 2}px`,
+          top: `${rect.height + 20}px`,
+          background: '#4af',
+          borderRadius: '2px',
+          padding: '2px 4px',
+          color: 'white',
+          fontSize: '13px',
+          whiteSpace: 'nowrap',
+          fontWeight: 'bold',
+          willChange: 'transform',
+          transform: 'translate(-50%, 0px)',
+          zIndex: 100,
+        }}
+      >
+        {Math.round(rect.offsetWidth)} x {Math.round(rect.offsetHeight)}
+      </div>
+    );
+  },
+};
