+++ public/app/features/canvas/runtime/scene.tsx
@@ -1,7 +1,6 @@
 import { css } from '@emotion/css';
 import Moveable from 'moveable';
 import React, { CSSProperties } from 'react';
-import { MoveableManagerInterface, Renderer } from 'react-moveable';
 import { ReplaySubject, Subject } from 'rxjs';
 import { first } from 'rxjs/operators';
 import Selecto from 'selecto';
@@ -27,8 +26,10 @@ import {
 } from 'app/features/dimensions/utils';
 import { LayerActionID } from 'app/plugins/panel/canvas/types';
 
-import { HorizontalConstraint, Placement, VerticalConstraint } from '../types';
+import { Placement } from '../types';
 
+import { constraintViewable } from './ConstraintViewable';
+import { dimensionViewable } from './DimensionViewable';
 import { ElementState } from './element';
 import { FrameState } from './frame';
 import { RootElement } from './root';
@@ -37,169 +38,6 @@ export interface SelectionParams {
   targets: Array<HTMLElement | SVGElement>;
   frame?: FrameState;
 }
-
-const constraintViewable = (scene: Scene) => ({
-  name: 'constraintViewable',
-  props: {},
-  events: {},
-  render(moveable: MoveableManagerInterface<any, any>, React: Renderer) {
-    const rect = moveable.getRect();
-    const targetElement = scene.findElementByTarget(moveable.state.target);
-
-    // If target is currently in motion or selection is more than 1 element don't display constraint visualizations
-    if (
-      targetElement?.isMoving ||
-      (scene.selecto?.getSelectedTargets() && scene.selecto?.getSelectedTargets().length > 1)
-    ) {
-      return;
-    }
-
-    let verticalConstraintVisualization = null;
-    let horizontalConstraintVisualization = null;
-
-    const constraint = targetElement?.options.constraint ?? {};
-
-    const borderStyle = '1px dashed #4af';
-
-    const centerIndicatorLineOne = React.createElement('div', {
-      style: {
-        position: 'absolute',
-        left: `${rect.width / 2}px`,
-        top: `${rect.height / 2 - rect.height / 16}px`,
-        borderLeft: borderStyle,
-        height: `${rect.height / 8}px`,
-        transform: 'rotate(45deg)',
-      },
-    });
-
-    const centerIndicatorLineTwo = React.createElement('div', {
-      style: {
-        position: 'absolute',
-        left: `${rect.width / 2}px`,
-        top: `${rect.height / 2 - rect.height / 16}px`,
-        borderLeft: borderStyle,
-        height: `${rect.height / 8}px`,
-        transform: 'rotate(-45deg)',
-      },
-    });
-
-    const centerIndicator = React.createElement('div', {}, [centerIndicatorLineOne, centerIndicatorLineTwo]);
-
-    const verticalConstraintTop = React.createElement('div', {
-      style: {
-        position: 'absolute',
-        left: `${rect.width / 2}px`,
-        bottom: '0px',
-        borderLeft: borderStyle,
-        height: '100vh',
-      },
-    });
-
-    const verticalConstraintBottom = React.createElement('div', {
-      style: {
-        position: 'absolute',
-        left: `${rect.width / 2}px`,
-        top: `${rect.height}px`,
-        borderLeft: borderStyle,
-        height: '100vh',
-      },
-    });
-
-    const verticalConstraintTopBottom = React.createElement('div', {}, [
-      verticalConstraintTop,
-      verticalConstraintBottom,
-    ]);
-
-    const verticalConstraintCenterLine = React.createElement('div', {
-      style: {
-        position: 'absolute',
-        left: `${rect.width / 2}px`,
-        top: `${rect.height / 4}px`,
-        borderLeft: borderStyle,
-        height: `${rect.height / 2}px`,
-      },
-    });
-
-    const verticalConstraintCenter = React.createElement('div', {}, [verticalConstraintCenterLine, centerIndicator]);
-
-    switch (constraint.vertical) {
-      case VerticalConstraint.Top:
-        verticalConstraintVisualization = verticalConstraintTop;
-        break;
-      case VerticalConstraint.Bottom:
-        verticalConstraintVisualization = verticalConstraintBottom;
-        break;
-      case VerticalConstraint.TopBottom:
-        verticalConstraintVisualization = verticalConstraintTopBottom;
-        break;
-      case VerticalConstraint.Center:
-        verticalConstraintVisualization = verticalConstraintCenter;
-        break;
-    }
-
-    const horizontalConstraintLeft = React.createElement('div', {
-      style: {
-        position: 'absolute',
-        right: '0px',
-        top: `${rect.height / 2}px`,
-        borderTop: borderStyle,
-        width: '100vw',
-      },
-    });
-
-    const horizontalConstraintRight = React.createElement('div', {
-      style: {
-        position: 'absolute',
-        left: `${rect.width}px`,
-        top: `${rect.height / 2}px`,
-        borderTop: borderStyle,
-        width: '100vw',
-      },
-    });
-
-    const horizontalConstraintLeftRight = React.createElement('div', {}, [
-      horizontalConstraintLeft,
-      horizontalConstraintRight,
-    ]);
-
-    const horizontalConstraintCenterLine = React.createElement('div', {
-      style: {
-        position: 'absolute',
-        left: `${rect.width / 4}px`,
-        top: `${rect.height / 2}px`,
-        borderTop: borderStyle,
-        width: `${rect.width / 2}px`,
-      },
-    });
-
-    const horizontalConstraintCenter = React.createElement('div', {}, [
-      horizontalConstraintCenterLine,
-      centerIndicator,
-    ]);
-
-    switch (constraint.horizontal) {
-      case HorizontalConstraint.Left:
-        horizontalConstraintVisualization = horizontalConstraintLeft;
-        break;
-      case HorizontalConstraint.Right:
-        horizontalConstraintVisualization = horizontalConstraintRight;
-        break;
-      case HorizontalConstraint.LeftRight:
-        horizontalConstraintVisualization = horizontalConstraintLeftRight;
-        break;
-      case HorizontalConstraint.Center:
-        horizontalConstraintVisualization = horizontalConstraintCenter;
-        break;
-    }
-
-    const constraintVisualization = React.createElement('div', {}, [
-      verticalConstraintVisualization,
-      horizontalConstraintVisualization,
-    ]);
-
-    return constraintVisualization;
-  },
-});
 
 export class Scene {
   styles = getStyles(config.theme2);
@@ -224,39 +62,6 @@ export class Scene {
   constructor(cfg: CanvasFrameOptions, enableEditing: boolean, public onSave: (cfg: CanvasFrameOptions) => void) {
     this.root = this.load(cfg, enableEditing);
   }
-
-  // Move to a file outside of scene
-  dimensionViewable = {
-    name: 'dimensionViewable',
-    props: {},
-    events: {},
-    render(moveable: MoveableManagerInterface<any, any>, React: Renderer) {
-      const rect = moveable.getRect();
-      return (
-        <div
-          key={'dimension-viewable'}
-          className={'moveable-dimension'}
-          style={{
-            position: 'absolute',
-            left: `${rect.width / 2}px`,
-            top: `${rect.height + 20}px`,
-            background: '#4af',
-            borderRadius: '2px',
-            padding: '2px 4px',
-            color: 'white',
-            fontSize: '13px',
-            whiteSpace: 'nowrap',
-            fontWeight: 'bold',
-            willChange: 'transform',
-            transform: 'translate(-50%, 0px)',
-            zIndex: 100,
-          }}
-        >
-          {Math.round(rect.offsetWidth)} x {Math.round(rect.offsetHeight)}
-        </div>
-      );
-    },
-  };
 
   getNextElementName = (isFrame = false) => {
     const label = isFrame ? 'Frame' : 'Element';
@@ -504,7 +309,7 @@ export class Scene {
     this.moveable = new Moveable(this.div!, {
       draggable: allowChanges,
       resizable: allowChanges,
-      ables: [this.dimensionViewable, constraintViewable(this)],
+      ables: [dimensionViewable, constraintViewable(this)],
       props: {
         dimensionViewable: allowChanges,
         constraintViewable: allowChanges,
